Refactor

+++ PurchaseManagement/MVVM/Models/PurchaseStatistics.cs
@@ -9,9 +9,9 @@ namespace PurchaseManagement.MVVM.Models
         public int Id { get; set; }
         public int Purchase_Id { get; set; }
         public int PurchaseCount { get; set; }
-        public long TotalPrice { get; set; }
-        public long TotalQuantity { get; set; }
-        public PurchaseStatistics(int p_id, int p_count, long total_price, long total_quantity)
+        public double TotalPrice { get; set; }
+        public double TotalQuantity { get; set; }
+        public PurchaseStatistics(int p_id, int p_count, double total_price, double total_quantity)
         {
             Purchase_Id = p_id;
             PurchaseCount = p_count;
